cpsu-chatbot-webhook/requirements.txt
# pip install -r .\requirements.txt
# .\.venv\Scripts\activate

# Web Framework
fastapi==0.135.2
uvicorn[standard]==0.42.0

# LINE SDK
line-bot-sdk==3.21.0

# LangChain Core & Agents
langchain==1.2.13
langgraph==1.1.3

# LangChain Providers
langchain-ollama==1.0.1

# Vector DB
qdrant-client==1.8.0

# Redis (Async + Semantic Cache)
redis==7.4.0
langchain-redis==0.2.5

# Object Storage
minio==7.2.14

# Message Queue
aio-pika==9.6.2

# Other
python-dotenv==1.2.1